Dynamic Product file support
This feature resolves #56 and adds an improvement to carbon-feature-plugin in supporting
a dynamically generated product file for the reference of publish-product and generate-runtime maven goals.
A dynamically generated product file during build-time overcomes the issue of having to manually update a statically maintained
product file with version-bumps for a release-build, thus contributing to a fully-automated release process.

Adding changes to resolve #56

Adding changes to resolve #56

Adding changes to resolve #56

Adding changes to resolve #56

Adding changes to resolve #56

Adding changes to resolve #56